requirements.txt
Django==1.8
Markdown==2.6.2
MySQL-python==1.2.5
django-cors-headers==1.0.0
django-filter==0.9.2
djangorestframework==3.1.1
djangorestframework-bulk==0.2.1
